bug fix on api structure completed
