+++ Experion.PickMyBook.Business/Service/IService/IBorrowingService.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Experion.PickMyBook.Business.Service.IService
+{
+    public interface IBorrowingService
+    {
+        Task<Borrowings> BorrowBookAsync(int userId, int bookId);
+        Task<Borrowings> UpdateBorrowingAsync(Borrowings borrowing);
+        Task<Borrowings> ReturnBookAsync(int userId, int bookId);
+    }
+}
